refactor LeafDbException, moved to model
removes weird direct dependency from API --> Services.
migrate LeafDbException.ThrowFrom to SqlException.MapThrow.
