+++ src/server/API/Controllers/Admin/AdminConceptEventController.cs
@@ -14,6 +14,7 @@ using Microsoft.Extensions.Logging;
 using Model.Authorization;
 using Services.Admin;
 using Model.Admin;
+using Model.Validation;
 using Services;
 using API.DTO.Admin;
