+++ src/server/API/Controllers/Admin/AdminSpecializationGroupController.cs
@@ -12,6 +12,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.Extensions.Logging;
 using Model.Authorization;
+using Model.Validation;
 using Services.Admin;
 using Model.Admin;
 using API.DTO.Admin;
