+++ src/server/Services/Admin/AdminConceptService.cs
@@ -14,11 +14,11 @@ using Model.Options;
 using System.Data.SqlClient;
 using System.Data;
 using Dapper;
-using Services.Authorization;
 using Services.Tables;
 using Services.Compiler;
 using System.Linq;
 using Model.Authorization;
+using Services.Extensions;
 
 namespace Services.Admin
 {
@@ -84,7 +84,7 @@ namespace Services.Admin
                 catch (SqlException se)
                 {
                     logger.LogError("Could not create concept. Concept:{@Concept} Code:{Code} Error:{Error}", c, se.ErrorCode, se.Message);
-                    LeafDbException.ThrowFrom(se);
+                    se.MapThrow();
                     throw;
                 }
             }
@@ -111,7 +111,7 @@ namespace Services.Admin
                 catch (SqlException se)
                 {
                     logger.LogError("Could not delete concept. Id:{Id} Code:{Code} Error:{Error}", id, se.ErrorCode, se.Message);
-                    LeafDbException.ThrowFrom(se);
+                    se.MapThrow();
                     throw;
                 }
             }
@@ -182,7 +182,7 @@ namespace Services.Admin
                 catch (SqlException se)
                 {
                     logger.LogError("Could not update concept. Concept:{@Concept} Code:{Code} Error:{Error}", c, se.ErrorCode, se.Message);
-                    LeafDbException.ThrowFrom(se);
+                    se.MapThrow();
                     throw;
                 }
             }
