+++ src/server/Services/Admin/AdminSpecializationGroupService.cs
@@ -14,7 +14,7 @@ using Model.Options;
 using System.Data.SqlClient;
 using System.Data;
 using Dapper;
-using Services.Authorization;
+using Services.Extensions;
 using Services.Tables;
 using System.Linq;
 using Model.Authorization;
@@ -62,7 +62,7 @@ namespace Services.Admin
                 catch (SqlException se)
                 {
                     logger.LogError("Could not create SpecializationGroup. SpecializationGroup:{@SpecializationGroup} Code:{Code} Error:{Error}", g, se.ErrorCode, se.Message);
-                    LeafDbException.ThrowFrom(se);
+                    se.MapThrow();
                     throw;
                 }
             }
@@ -88,7 +88,7 @@ namespace Services.Admin
                 catch (SqlException se)
                 {
                     logger.LogError("Could not delete SpecializationGroup. Id:{Id} Code:{Code} Error:{Error}", id, se.ErrorCode, se.Message);
-                    LeafDbException.ThrowFrom(se);
+                    se.MapThrow();
                     throw;
                 }
             }
@@ -140,7 +140,7 @@ namespace Services.Admin
                 catch (SqlException se)
                 {
                     logger.LogInformation("Could not update SpecializationGroup. SpecializationGroup:{@SpecializationGroup} Code:{Code} Error:{Error}", record, se.ErrorCode, se.Message);
-                    LeafDbException.ThrowFrom(se);
+                    se.MapThrow();
                     throw;
                 }
             }
