+++ src/server/Services/Tables/ISqlTableType.cs
@@ -5,7 +5,7 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 using System;
 using System.Data;
-namespace Services
+namespace Services.Tables
 {
     public interface ISqlTableType
     {
